Ticket 1093: moved custom instrument info ui classes to new package
