+++ base/uk.ac.stfc.isis.ibex.ui.mainmenu/src/uk/ac/stfc/isis/ibex/ui/mainmenu/instrument/custom/CustomInstrumentConfigDialog.java
@@ -17,7 +17,7 @@
  * http://opensource.org/licenses/eclipse-1.0.php
  */
 
-package uk.ac.stfc.isis.ibex.ui.mainmenu.instrument;
+package uk.ac.stfc.isis.ibex.ui.mainmenu.instrument.custom;
 
 import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
